Fix Delete Customer

+++ BLL/Services/CustomerService.cs
@@ -1,5 +1,6 @@
 ﻿using DAL.Models.Difinitions;
 using DAL.Repositories.Abstractions;
+using Microsoft.IdentityModel.Tokens;
 
 namespace BLL.Services
 {
@@ -41,11 +42,35 @@ namespace BLL.Services
         public async Task DeleteAsync(int id)
         {
             var entity = await _repository.GetByIdAsync(id);
-            if (entity != null)
+            if (entity != null && entity.Warehouses.IsNullOrEmpty() && entity.SubCustomers.IsNullOrEmpty())
             {
                 _repository.Delete(entity);
                 await _repository.SaveChangesAsync();
             }
+        }
+
+        public async Task<bool> DeleteAsyncWithCheck(int id)
+        {
+            var entity = await _repository.GetByIdAsync(id);
+            if (entity == null)
+                return false;
+
+
+            if (!entity.Warehouses.IsNullOrEmpty() || !entity.SubCustomers.IsNullOrEmpty())
+            {
+                return false;
+            }
+
+            //if (entity.Warehouses.Any() || entity.SubCustomers.Any())
+            //{
+            //    return false;
+            //}
+
+
+
+            _repository.Delete(entity);
+            await _repository.SaveChangesAsync();
+            return true;
         }
     }
